backend/requirements.txt
fastapi
uvicorn[standard]
faster-whisper
python-multipart
# pydub is no longer strictly needed for splitting if we trust faster-whisper, 
# but keeping it just in case we need simple audio manipulation later is fine, 
# though we want to minimize image size. Let's keep it minimal.
# pydub
numpy
torch
torchaudio
pyannote.audio
python-dotenv
